reduce java's latency by disabling statistics
Latency measurements are hard.

All of our drivers print some stats in the main thread, that wasn't a
big problem for any language. However, it seems to be for some Java
Garbage Collectors. Disabling these prints reduced latency for the
Shenandoah, Epsilon and ZGC collectors from ~300 µs to ~40-100 µs.

We didn't catch that earlier because the other GCs are either not
affected by this or have a *higher* latency when doing less by not
printing statistics. For example, G1's latency increases by 400 µs
when you disable printing statistics...